Add Makefile, Dockerfile, notebooks and the paper; clean lint and types
`make all` is the definition of done and the Dockerfile makes it reproducible;
TeX Live is installed at build time so a container run needs no network.

paper/main.tex contains no typed numbers. Every quantitative claim is a macro
that hcmtwin.report generates from the result CSVs at build time, so the paper
cannot describe a run other than the one it ships with, and a missing stage
renders as a visible "??" rather than compiling with a silent gap.

Two real defects found while wiring this up, both in what the inference was
allowed to learn from:

  - Heart rate was in the likelihood. It is a genuine observable, but in this
    model it is an input the analyst knows exactly: identical for every candidate
    parameter vector, so no posterior can learn from it. It also broke the
    surrogate's diagnostics, which is how it surfaced. A coefficient of
    determination divides by the output's variance, and for a constant output
    that is zero, so forty emulators reported an R-squared of minus two hundred
    thousand that was floating-point noise divided by nothing.
  - Left-ventricular mass was in the likelihood for the same reason, and it is
    the more interesting case. It is the measurement that defines the disease on
    imaging, but wall volume is pinned as known and mass is wall volume times a
    density constant, so conditional on the geometry it is already determined.
    That is not a defect of the measurement; it is the point of separating shape
    from material.

Constant outputs now report an undefined R-squared instead of a meaningless one,
and are counted separately.

Also: ruff and mypy clean across 29 modules. The type fixes are real rather than
suppressions. The Backend protocol claimed to preserve a type variable that
neither concrete backend can honour (one is float-only, the other array-only),
and several layer functions were annotated float where the cohort path genuinely
passes arrays.

+++ src/hcmtwin/circulation.py
@@ -26,7 +26,7 @@ from .backend import SCALAR, Backend, Numeric
 
 
 def venous_pressure_mmhg(
-    total_volume_ml: float,
+    total_volume_ml: Numeric,
     cavity_volume_ml: Numeric,
     arterial_pressure_mmhg: Numeric,
     c_art_ml_per_mmhg: float,
@@ -60,7 +60,7 @@ def valve_flow_ml_per_s(
 def systemic_flow_ml_per_s(
     arterial_pressure_mmhg: Numeric,
     venous_pressure_mmhg_value: Numeric,
-    r_sys_mmhg_s_per_ml: float,
+    r_sys_mmhg_s_per_ml: Numeric,
 ) -> Numeric:
     """Flow through the systemic bed, mL/s. Bidirectional: no valve here."""
     return (arterial_pressure_mmhg - venous_pressure_mmhg_value) / r_sys_mmhg_s_per_ml
